Set UUID version bits according to v4 (pseudorandom)
RFC 4122 describes UUID versioning (i.e. identifying how the UUID was
generated) by setting some of the bits. See the table in section 4.1.3:

  https://tools.ietf.org/html/rfc4122#section-4.1.3

The best match for us is v4, i.e. pseudorandom generator. The other
generators are substantially different (use MAC address or namespace,
work with very different timestamps or hashing schemes), so it's best
to use v4 even if the prefix is not properly random. We also set the
"variant" as specified by the RFC.

In total this sets 7 bits, so it reduces the randomness a bit. Compared
to the prefix it's much smaller, and it's not specific to this generator
(any UUID generator should be doing this).

Reported-by: sehrope