Rework prefix calculation and copying
The code attempted to copy prefixes or arbitrary length (number of bits
used to encode the block number), but combining that with the random
part in a way that would sort correctly seems harder than expected.

It might be possible to fix that, but it seems easier to ensure the
prefix is always a multiple of whole bytes. As this adjustment happens
automatically, the actual number of blocks may be much higher than what
the user requested. To deal with that, we recalculate the block size
(values to consider as one block), to keep the same number of values
before the UUID wraps.

For example if the user specifies block_size=1000 and block_count=1000,
we adjust block_count=65536 (so prefix is 2B) and then adjust the block
size to 16. This means the generated UUID values will wrap after about
1M values in both cases, but there will be more (smaller) blocks.